eina_hash: fix possible issue
As seen in ed84d7d9007b297c00540f38bc07422a3be7267c, values can
overflow.

This here is a signed integer, and the hash can come from a user and
gets right shifted by a value between 2 and 17 bits, this could result
in a maximum big integer in one hash head, but a 0 in the other hash
head. This means the operation:

left->hash - right->hash could overflow.

This can be worked arround with changing
(left->hash - right->hash < 0)
to
left->hash < right->hash

This is not overflowting,because jmg can handle two 32bit words